r-car m2: support Bluetooth hardware
In order to have Bluetooth hardware detection, and to make
BlueZ commands (such as "hciconfig", "hcitool"...) work,
we need to enable kernel functionalities. Some of them are
designed to be run as modules.

We also make sure that all images include them, since it
does not depend on a specific board's hardware.

PS : we will only build them if "bluetooth" has been
specified in DISTRO_FEATURES, hence saving time.

Change-Id: I14e1bdd7e4e63a7484bb1516dda601e3c03fd58f
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>