r-car m2: add EGL driver EGL_NO_DISPLAY workaroud
Weston 1.8.0 uses a new method of detection for client
EGL extensions, which is incompatible with the driver
provided by meta-renesas. We revert back to the old
behavior as it is not strictly necessary for rendering.

This allows Weston 1.8.0 to start in DRM mode on
Porter/Koelsch with latest AGL base.

Signed-off-by: Manuel Bachmann <mbc@iot.bzh>
Change-Id: I0658b3beff96c7f173dbb4f85e8f18d3fce768c1