r-car m2: add a "libGLESv2.so.2" symbolic link for software
Some popular software, including Chromium/QtWebEngine, look
for a library named "libGLESv2.so.2" on the filesystem to
initialize their OpenGL ES backend :

https://github.com/qtproject/qtwebengine/blob/dev/src/core/
surface_factory_qt.cpp#L83

but the proprietary userspace driver is named "libGLESv2.so"
instead. Let us create a symbolic link which will satisfy
this requirement (we do not rename the binary and make
"libGLESv2.so" a symbolic link, because this would trigger
a Yocto QA warning and prevent building).

Change-Id: Iea901694e90cfd4a786087507cfc77302c9d3b8a
Signed-off-by: Manuel Bachmann <mbc@iot.bzh>
Author: Yannick Gicquel <yannick.gicquel@iot.bzh>
Author: Manuel Bachmann <mbc@iot.bzh>